feat: return onCloseDialog function in useDialog hook (#28)
