+++ src/dialog/Dialog.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/prop-types */
-import React, { useState } from 'react';
+import React, { useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 
 import { useOpenState } from '../utils/useOpenState';
@@ -8,22 +8,30 @@ import { safeClick } from '../utils/event-handlers';
 
 import { ModalPortal } from './ModalPortal';
 import { Modal } from './Modal';
+import { hideModal } from './helpers';
 
 export function useDialog({ onlyRenderContentIfIsOpen = true, ...props }) {
   const { isOpen, open, close } = useOpenState();
   const [dialogBodyProps, setDialogBodyProps] = useState({});
+  const modalRef = useRef(null);
+
+  const onClose = () => {
+    hideModal(modalRef);
+    close();
+  };
 
   return {
     showDialog(_props) {
       setDialogBodyProps(_props);
       open();
     },
+    closeDialog: onClose,
     DialogPortal() {
       return onlyRenderContentIfIsOpen && !isOpen() ? (
         <></>
       ) : (
         <ModalPortal isOpen={isOpen()} title={props.title}>
-          <Modal {...props} dialogBodyProps={dialogBodyProps} onClose={close} isOpen={isOpen()} />
+          <Modal {...props} modalRef={modalRef} dialogBodyProps={dialogBodyProps} onClose={onClose} isOpen={isOpen()} />
         </ModalPortal>
       );
     },
@@ -33,6 +41,13 @@ export function useDialog({ onlyRenderContentIfIsOpen = true, ...props }) {
 export function Dialog({ children, onlyRenderContentIfIsOpen, ...props }) {
   const { isOpen, open, close } = useOpenState();
 
+  const modalRef = useRef(null);
+
+  const onClose = () => {
+    hideModal(modalRef);
+    close();
+  };
+
   return (
     <React.Fragment>
       <DialogTrigger open={open}>{children}</DialogTrigger>
@@ -41,7 +56,7 @@ export function Dialog({ children, onlyRenderContentIfIsOpen, ...props }) {
         <></>
       ) : (
         <ModalPortal isOpen={isOpen()} title={props.title}>
-          <Modal {...props} onClose={close} isOpen={isOpen()} />
+          <Modal {...props} modalRef={modalRef} onClose={onClose} isOpen={isOpen()} />
         </ModalPortal>
       )}
     </React.Fragment>
